+++ module/apps/coc7-chat-message.js
@@ -55,23 +55,56 @@ export class CoC7ChatMessage {
       [CoC7ChatMessage.CARD_TYPE_OPPOSED]: 'CoC7.OpposedRollCard'
       // [CoC7ChatMessage.CARD_TYPE_GROUP]: 'CoC7.GroupRollCard' - NYI
     }
-    if (config.rollType === CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE && config.attribute === 'san') {
+    if (
+      config.rollType === CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE &&
+      config.attribute === 'san'
+    ) {
       select[CoC7ChatMessage.CARD_TYPE_SAN_CHECK] = 'CoC7.SanityLossEncounter'
     }
     return select
   }
 
   static normalizeRequest (options) {
-    if (typeof options.event === 'undefined' || typeof options.cardType === 'undefined' || typeof options.rollType === 'undefined') {
-      ui.notifications.error(game.i18n.format('CoC7.ErrorNotFound', { value: game.i18n.localize('CoC7.Roll') }))
+    if (
+      typeof options.event === 'undefined' ||
+      typeof options.cardType === 'undefined' ||
+      typeof options.rollType === 'undefined'
+    ) {
+      ui.notifications.error(
+        game.i18n.format('CoC7.ErrorNotFound', {
+          value: game.i18n.localize('CoC7.Roll')
+        })
+      )
       return false
     }
-    if (![CoC7ChatMessage.CARD_TYPE_COMBINED, CoC7ChatMessage.CARD_TYPE_GROUP, CoC7ChatMessage.CARD_TYPE_NORMAL, CoC7ChatMessage.CARD_TYPE_OPPOSED, CoC7ChatMessage.CARD_TYPE_SAN_CHECK].includes(options.cardType)) {
-      ui.notifications.error(game.i18n.format('CoC7.ErrorInvalidFormula', { value: game.i18n.localize('CoC7.ErrorInvalidCardType') }))
+    if (
+      ![
+        CoC7ChatMessage.CARD_TYPE_COMBINED,
+        CoC7ChatMessage.CARD_TYPE_GROUP,
+        CoC7ChatMessage.CARD_TYPE_NORMAL,
+        CoC7ChatMessage.CARD_TYPE_OPPOSED,
+        CoC7ChatMessage.CARD_TYPE_SAN_CHECK
+      ].includes(options.cardType)
+    ) {
+      ui.notifications.error(
+        game.i18n.format('CoC7.ErrorInvalidFormula', {
+          value: game.i18n.localize('CoC7.ErrorInvalidCardType')
+        })
+      )
       return false
     }
-    if (![CoC7ChatMessage.ROLL_TYPE_SKILL, CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC, CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE].includes(options.rollType)) {
-      ui.notifications.error(game.i18n.format('CoC7.ErrorInvalidFormula', { value: game.i18n.localize('CoC7.ErrorInvalidRollType') }))
+    if (
+      ![
+        CoC7ChatMessage.ROLL_TYPE_SKILL,
+        CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC,
+        CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE
+      ].includes(options.rollType)
+    ) {
+      ui.notifications.error(
+        game.i18n.format('CoC7.ErrorInvalidFormula', {
+          value: game.i18n.localize('CoC7.ErrorInvalidRollType')
+        })
+      )
       return false
     }
     const config = {
@@ -92,16 +125,26 @@ export class CoC7ChatMessage {
       case CoC7ChatMessage.ROLL_TYPE_SKILL:
       case CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC:
       case CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE:
-        config.options.skillId = options.event.currentTarget.closest('.item')?.dataset.skillId
-        config.options.itemId = options.event.currentTarget.closest('.item')?.dataset.itemId
-        config.options.characteristic = options.event.currentTarget.parentElement.dataset.characteristic
-        config.options.attribute = options.event.currentTarget.parentElement.dataset.attrib
+        config.options.skillId =
+          options.event.currentTarget.closest('.item')?.dataset.skillId
+        config.options.itemId =
+          options.event.currentTarget.closest('.item')?.dataset.itemId
+        config.options.characteristic =
+          options.event.currentTarget.parentElement.dataset.characteristic
+        config.options.attribute =
+          options.event.currentTarget.parentElement.dataset.attrib
         config.dialogOptions.attribute = config.options.attribute
-        config.options.actorId = options.event.currentTarget.closest('form').dataset.actorId
-        config.options.tokenKey = options.event.currentTarget.closest('form').dataset.tokenId
-        config.options.weaponAltSkill = options.event.currentTarget.classList.contains('alternativ-skill')
+        config.options.actorId =
+          options.event.currentTarget.closest('form').dataset.actorId
+        config.options.tokenKey =
+          options.event.currentTarget.closest('form').dataset.tokenId
+        config.options.weaponAltSkill =
+          options.event.currentTarget.classList.contains('alternativ-skill')
         config.options.actor = options.actor
-        if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_SKILL && config.options.isCombat) {
+        if (
+          config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_SKILL &&
+          config.options.isCombat
+        ) {
           const item = config.options.actor.items.get(config.options.itemId)
           if (!item) {
             return false
@@ -111,45 +154,97 @@ export class CoC7ChatMessage {
             config.dialogOptions.rollType = CoC7ChatMessage.ROLL_TYPE_COMBAT
           }
         }
-        if (typeof config.options.actorId === 'undefined' || typeof config.options.tokenKey === 'undefined' || typeof config.options.actor === 'undefined') {
-          ui.notifications.error(game.i18n.format('CoC7.ErrorInvalidFormula', { value: game.i18n.localize('CoC7.ErrorActor') }))
+        if (
+          typeof config.options.actorId === 'undefined' ||
+          typeof config.options.tokenKey === 'undefined' ||
+          typeof config.options.actor === 'undefined'
+        ) {
+          ui.notifications.error(
+            game.i18n.format('CoC7.ErrorInvalidFormula', {
+              value: game.i18n.localize('CoC7.ErrorActor')
+            })
+          )
           return false
         }
-        if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_SKILL && typeof config.options.skillId === 'undefined') {
-          ui.notifications.error(game.i18n.format('CoC7.ErrorInvalidFormula', { value: game.i18n.localize('CoC7.Entities.Skill') }))
+        if (
+          config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_SKILL &&
+          typeof config.options.skillId === 'undefined'
+        ) {
+          ui.notifications.error(
+            game.i18n.format('CoC7.ErrorInvalidFormula', {
+              value: game.i18n.localize('CoC7.Entities.Skill')
+            })
+          )
           return false
-        } else if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC) {
+        } else if (
+          config.dialogOptions.rollType ===
+          CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC
+        ) {
           if (typeof config.options.characteristic === 'undefined') {
-            ui.notifications.error(game.i18n.format('CoC7.ErrorInvalidFormula', { value: game.i18n.localize('CoC7.Characteristic') }))
+            ui.notifications.error(
+              game.i18n.format('CoC7.ErrorInvalidFormula', {
+                value: game.i18n.localize('CoC7.Characteristic')
+              })
+            )
             return false
-          } else if (!game.user.isGM && !config.options.actor.data.data.characteristics[config.options.characteristic]?.value) {
+          } else if (
+            !game.user.isGM &&
+            !config.options.actor.data.data.characteristics[
+              config.options.characteristic
+            ]?.value
+          ) {
             return false
           }
-        } else if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE) {
+        } else if (
+          config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE
+        ) {
           if (typeof config.options.attribute === 'undefined') {
-            ui.notifications.error(game.i18n.format('CoC7.ErrorInvalidFormula', { value: game.i18n.localize('CoC7.Attribute') }))
+            ui.notifications.error(
+              game.i18n.format('CoC7.ErrorInvalidFormula', {
+                value: game.i18n.localize('CoC7.Attribute')
+              })
+            )
             return false
-          } else if (!game.user.isGM && !config.options.actor.data.data.attribs[config.options.attribute]?.value) {
+          } else if (
+            !game.user.isGM &&
+            !config.options.actor.data.data.attribs[config.options.attribute]
+              ?.value
+          ) {
             return false
           }
-        } else if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_COMBAT) {
+        } else if (
+          config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_COMBAT
+        ) {
           if (typeof config.options.itemId === 'undefined') {
-            ui.notifications.error(game.i18n.format('CoC7.ErrorInvalidFormula', { value: game.i18n.localize('CoC7.Entities.Weapon') }))
+            ui.notifications.error(
+              game.i18n.format('CoC7.ErrorInvalidFormula', {
+                value: game.i18n.localize('CoC7.Entities.Weapon')
+              })
+            )
             return false
           } else if (config.options.altKey) {
             return false
           }
         }
-        config.options.hasPlayerOwner = (config.options.actor.hasPlayerOwner ?? false)
-        if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC) {
-          config.dialogOptions.disableFlatThresholdModifier = config.options.isCtrlKey
-          config.dialogOptions.disableFlatDiceModifier = config.options.isCtrlKey
+        config.options.hasPlayerOwner =
+          config.options.actor.hasPlayerOwner ?? false
+        if (
+          config.dialogOptions.rollType ===
+          CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC
+        ) {
+          config.dialogOptions.disableFlatThresholdModifier =
+            config.options.isCtrlKey
+          config.dialogOptions.disableFlatDiceModifier =
+            config.options.isCtrlKey
         } else {
           config.dialogOptions.disableFlatThresholdModifier = false
           config.dialogOptions.disableFlatDiceModifier = false
         }
         config.dialogOptions.modifier = 0
-        config.dialogOptions.difficulty = CoC7Check.difficultyLevel[game.settings.get('CoC7', 'defaultCheckDifficulty')]
+        config.dialogOptions.difficulty =
+          CoC7Check.difficultyLevel[
+            game.settings.get('CoC7', 'defaultCheckDifficulty')
+          ]
         config.dialogOptions.flatDiceModifier = 0
         config.dialogOptions.flatThresholdModifier = 0
         break
@@ -162,7 +257,14 @@ export class CoC7ChatMessage {
     if (config === false) {
       return
     }
-    if (config.options.isCtrlKey && game.user.isGM && [CoC7ChatMessage.CARD_TYPE_NORMAL, CoC7ChatMessage.CARD_TYPE_SAN_CHECK].includes(config.dialogOptions.cardType)) {
+    if (
+      config.options.isCtrlKey &&
+      game.user.isGM &&
+      [
+        CoC7ChatMessage.CARD_TYPE_NORMAL,
+        CoC7ChatMessage.CARD_TYPE_SAN_CHECK
+      ].includes(config.dialogOptions.cardType)
+    ) {
       CoC7ChatMessage.createLink(config)
     } else {
       if (!config.options.shiftKey) {
@@ -187,21 +289,33 @@ export class CoC7ChatMessage {
             actorKey: config.options.actor.actorKey,
             forceModifiers: config.options.shiftKey
           }
-          if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_SKILL) {
+          if (
+            config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_SKILL
+          ) {
             linkData.type = 'skill'
-            linkData.name = config.options.actor.items.get(config.options.skillId)?.shortName
+            linkData.name = config.options.actor.items.get(
+              config.options.skillId
+            )?.shortName
             if (!linkData.name) return
-          } else if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC) {
+          } else if (
+            config.dialogOptions.rollType ===
+            CoC7ChatMessage.ROLL_TYPE_CHARACTERISTIC
+          ) {
             linkData.type = 'characteristic'
             linkData.name = config.options.characteristic
-          } else if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE) {
+          } else if (
+            config.dialogOptions.rollType ===
+            CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE
+          ) {
             if (config.options.altKey && config.options.attribute === 'san') {
               linkData.check = 'sanloss'
             } else {
               linkData.type = 'attribute'
               linkData.name = config.options.attribute
             }
-          } else if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_COMBAT) {
+          } else if (
+            config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_COMBAT
+          ) {
             linkData.check = 'item'
             linkData.type = 'weapon'
             linkData.name = config.options.weaponName
@@ -228,8 +342,12 @@ export class CoC7ChatMessage {
       }
       config.dialogOptions.modifier = Number(usage.get('bonusDice'))
       config.dialogOptions.difficulty = Number(usage.get('difficulty'))
-      config.dialogOptions.flatDiceModifier = Number(usage.get('flatDiceModifier'))
-      config.dialogOptions.flatThresholdModifier = Number(usage.get('flatThresholdModifier'))
+      config.dialogOptions.flatDiceModifier = Number(
+        usage.get('flatDiceModifier')
+      )
+      config.dialogOptions.flatThresholdModifier = Number(
+        usage.get('flatThresholdModifier')
+      )
     }
   }
 
@@ -265,17 +383,25 @@ export class CoC7ChatMessage {
           const check = new CoC7Check()
           check.diceModifier = config.dialogOptions.modifier
           check.difficulty = config.dialogOptions.difficulty
-          check.actor = !config.options.tokenKey ? config.options.actorId : config.options.tokenKey
+          check.actor = !config.options.tokenKey
+            ? config.options.actorId
+            : config.options.tokenKey
           check.flatDiceModifier = config.dialogOptions.flatDiceModifier
-          check.flatThresholdModifier = config.dialogOptions.flatThresholdModifier
+          check.flatThresholdModifier =
+            config.dialogOptions.flatThresholdModifier
           check.standby =
             game.settings.get('CoC7', 'stanbyGMRolls') &&
             game.user.isGM &&
             config.options.hasPlayerOwner
-          if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_SKILL) {
+          if (
+            config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_SKILL
+          ) {
             check.skill = config.options.skillId
             await check.roll()
-          } else if (config.dialogOptions.rollType === CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE) {
+          } else if (
+            config.dialogOptions.rollType ===
+            CoC7ChatMessage.ROLL_TYPE_ATTRIBUTE
+          ) {
             await check.rollAttribute(config.options.attribute)
           } else {
             await check.rollCharacteristic(config.options.characteristic)
@@ -287,14 +413,18 @@ export class CoC7ChatMessage {
       case CoC7ChatMessage.CARD_TYPE_COMBINED:
         {
           const check = new CoC7Check()
-          check.actor = !config.options.tokenKey ? config.options.actorId : config.options.tokenKey
+          check.actor = !config.options.tokenKey
+            ? config.options.actorId
+            : config.options.tokenKey
           check.characteristic = config.options.characteristic
           check.attribute = config.options.attribute
           check.skillId = config.options.skillId
           check.rollMode = game.settings.get('core', 'rollMode')
           check.initiator = game.user.id
           let data = {}
-          if (config.dialogOptions.cardType === CoC7ChatMessage.CARD_TYPE_OPPOSED) {
+          if (
+            config.dialogOptions.cardType === CoC7ChatMessage.CARD_TYPE_OPPOSED
+          ) {
             data = {
               type: OpposedCheckCard.defaultConfig.type,
               combat: config.options.isCombat,
@@ -313,7 +443,9 @@ export class CoC7ChatMessage {
           }
           data.roll = check.JSONRollData
           data._rollMode = game.settings.get('core', 'rollMode')
-          if (config.dialogOptions.cardType === CoC7ChatMessage.CARD_TYPE_OPPOSED) {
+          if (
+            config.dialogOptions.cardType === CoC7ChatMessage.CARD_TYPE_OPPOSED
+          ) {
             OpposedCheckCard.dispatch(data)
           } else {
             CombinedCheckCard.dispatch(data)